Flex

+++ simulation/src/point-to-point/model/cmsketch.cc
@@ -31,12 +31,12 @@ namespace ns3
             fprintf(stderr, "Unable to initialize the count-min sketch since both error_rate and confidence must be positive!\n");
             return CMS_ERROR;
         }
-        uint32_t width = ceil(2 / error_rate);
-        uint32_t depth = ceil((-1 * log(1 - confidence)) / LOG_TWO);
+        int32_t width = ceil(2 / error_rate);
+        int32_t depth = ceil((-1 * log(1 - confidence)) / LOG_TWO);
         return __setup_cms(cms, width, depth, error_rate, confidence, hash_function);
     }
 
-    int CountMinSketch::cms_init_alt(CountMinSketch* cms, uint32_t width, uint32_t depth, cms_hash_function hash_function) {
+    int CountMinSketch::cms_init_alt(CountMinSketch* cms, int32_t width, int32_t depth, cms_hash_function hash_function) {
         if (depth < 1 || width < 1) {
             fprintf(stderr, "Unable to initialize the count-min sketch since either width or depth is 0!\n");
             return CMS_ERROR;
@@ -60,7 +60,7 @@ namespace ns3
     }
 
     int CountMinSketch::cms_clear(CountMinSketch* cms) {
-        uint32_t i, j = cms->width * cms->depth;
+        uint64_t i, j = cms->width * cms->depth;
         for (i = 0; i < j; ++i) {
             cms->bins[i] = 0;
         }
@@ -68,7 +68,7 @@ namespace ns3
         return CMS_SUCCESS;
     }
 
-    int32_t CountMinSketch::cms_add_inc_alt(CountMinSketch* cms, uint64_t* hashes, unsigned int num_hashes, uint32_t x) {
+    uint64_t CountMinSketch::cms_add_inc_alt(CountMinSketch* cms, uint64_t* hashes, unsigned int num_hashes, uint64_t x) {
         if (num_hashes < cms->depth) {
             fprintf(stderr, "Insufficient hashes to complete the addition of the element to the count-min sketch!");
             return CMS_ERROR;
@@ -86,21 +86,21 @@ namespace ns3
         return num_add;
     }
 
-    int32_t CountMinSketch::cms_add_inc(CountMinSketch* cms, const char* key, unsigned int x) {
+    uint64_t CountMinSketch::cms_add_inc(CountMinSketch* cms, const char* key, uint64_t x) {
         uint64_t* hashes = cms_get_hashes(cms, key);
-        int32_t num_add = cms_add_inc_alt(cms, hashes, cms->depth, x);
+        uint64_t num_add = cms_add_inc_alt(cms, hashes, cms->depth, x);
         free(hashes);
         return num_add;
     }
 
-    int32_t CountMinSketch::cms_remove_inc_alt(CountMinSketch* cms, uint64_t* hashes, unsigned int num_hashes, unsigned int x) {
+    uint64_t CountMinSketch::cms_remove_inc_alt(CountMinSketch* cms, uint64_t* hashes, unsigned int num_hashes, uint64_t x) {
         if (num_hashes < cms->depth) {
             fprintf(stderr, "Insufficient hashes to complete the removal of the element to the count-min sketch!");
             return CMS_ERROR;
         }
-        int32_t num_add = INT32_MAX;
+        uint64_t num_add = INT32_MAX;
         for (unsigned int i = 0; i < cms->depth; ++i) {
-            uint32_t bin = (hashes[i] % cms->width) + (i * cms->width);
+            uint64_t bin = (hashes[i] % cms->width) + (i * cms->width);
             cms->bins[bin] = __safe_sub(cms->bins[bin], x);
             if (cms->bins[bin] < num_add) {
                 num_add = cms->bins[bin];
@@ -110,21 +110,21 @@ namespace ns3
         return num_add;
     }
 
-    int32_t CountMinSketch::cms_remove_inc(CountMinSketch* cms, const char* key, uint32_t x) {
+    uint64_t CountMinSketch::cms_remove_inc(CountMinSketch* cms, const char* key, uint64_t x) {
         uint64_t* hashes = cms_get_hashes(cms, key);
-        int32_t num_add = cms_remove_inc_alt(cms, hashes, cms->depth, x);
+        uint64_t num_add = cms_remove_inc_alt(cms, hashes, cms->depth, x);
         free(hashes);
         return num_add;
     }
 
-    int32_t CountMinSketch::cms_check_alt(CountMinSketch* cms, uint64_t* hashes, unsigned int num_hashes) {
+    uint64_t CountMinSketch::cms_check_alt(CountMinSketch* cms, uint64_t* hashes, unsigned int num_hashes) {
         if (num_hashes < cms->depth) {
             fprintf(stderr, "Insufficient hashes to complete the min lookup of the element to the count-min sketch!");
             return CMS_ERROR;
         }
-        int32_t num_add = INT32_MAX;
+        uint64_t num_add = INT32_MAX;
         for (unsigned int i = 0; i < cms->depth; ++i) {
-            uint32_t bin = (hashes[i] % cms->width) + (i * cms->width);
+            uint64_t bin = (hashes[i] % cms->width) + (i * cms->width);
             if (cms->bins[bin] < num_add) {
                 num_add = cms->bins[bin];
             }
@@ -132,48 +132,48 @@ namespace ns3
         return num_add;
     }
 
-    int32_t CountMinSketch::cms_check(CountMinSketch* cms, const char* key) {
+    uint64_t CountMinSketch::cms_check(CountMinSketch* cms, const char* key) {
         uint64_t* hashes = cms_get_hashes(cms, key);
-        int32_t num_add = cms_check_alt(cms, hashes, cms->depth);
+        uint64_t num_add = cms_check_alt(cms, hashes, cms->depth);
         free(hashes);
         return num_add;
     }
 
-    int32_t CountMinSketch::cms_check_mean_alt(CountMinSketch* cms, uint64_t* hashes, unsigned int num_hashes) {
+    uint64_t CountMinSketch::cms_check_mean_alt(CountMinSketch* cms, uint64_t* hashes, unsigned int num_hashes) {
         if (num_hashes < cms->depth) {
             fprintf(stderr, "Insufficient hashes to complete the mean lookup of the element to the count-min sketch!");
             return CMS_ERROR;
         }
-        int32_t num_add = 0;
+        uint64_t num_add = 0;
         for (unsigned int i = 0; i < cms->depth; ++i) {
-            uint32_t bin = (hashes[i] % cms->width) + (i * cms->width);
+            uint64_t bin = (hashes[i] % cms->width) + (i * cms->width);
             num_add += cms->bins[bin];
         }
         return num_add / cms->depth;
     }
 
-    int32_t CountMinSketch::cms_check_mean(CountMinSketch* cms, const char* key) {
+    uint64_t CountMinSketch::cms_check_mean(CountMinSketch* cms, const char* key) {
         uint64_t* hashes = cms_get_hashes(cms, key);
-        int32_t num_add = cms_check_mean_alt(cms, hashes, cms->depth);
+        uint64_t num_add = cms_check_mean_alt(cms, hashes, cms->depth);
         free(hashes);
         return num_add;
     }
 
-    int32_t CountMinSketch::cms_check_mean_min_alt(CountMinSketch* cms, uint64_t* hashes, unsigned int num_hashes) {
+    uint64_t CountMinSketch::cms_check_mean_min_alt(CountMinSketch* cms, uint64_t* hashes, unsigned int num_hashes) {
         if (num_hashes < cms->depth) {
             fprintf(stderr, "Insufficient hashes to complete the mean-min lookup of the element to the count-min sketch!");
             return CMS_ERROR;
         }
-        int32_t num_add = 0;
+        uint64_t num_add = 0;
         int64_t* mean_min_values = (int64_t*)calloc(cms->depth, sizeof(int64_t));
         for (unsigned int i = 0; i < cms->depth; ++i) {
-            uint32_t bin = (hashes[i] % cms->width) + (i * cms->width);
-            int32_t val = cms->bins[bin];
+            uint64_t bin = (hashes[i] % cms->width) + (i * cms->width);
+            uint64_t val = cms->bins[bin];
             mean_min_values[i] = val - ((cms->elements_added - val) / (cms->width - 1));
         }
         // return the median of the mean_min_value array... need to sort first
         qsort(mean_min_values, cms->depth, sizeof(int64_t), __compare);
-        int32_t n = cms->depth;
+        uint64_t n = cms->depth;
         if (n % 2 == 0) {
             num_add = (mean_min_values[n/2] + mean_min_values[n/2 - 1]) / 2;
         } else {
@@ -183,9 +183,9 @@ namespace ns3
         return num_add;
     }
 
-    int32_t CountMinSketch::cms_check_mean_min(CountMinSketch* cms, const char* key) {
+    uint64_t CountMinSketch::cms_check_mean_min(CountMinSketch* cms, const char* key) {
         uint64_t* hashes = cms_get_hashes(cms, key);
-        int32_t num_add = cms_check_mean_min_alt(cms, hashes, cms->depth);
+        uint64_t num_add = cms_check_mean_min_alt(cms, hashes, cms->depth);
         free(hashes);
         return num_add;
     }
